old/pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "markdown-to-html-converter"
version = "0.1.0"
description = "Convert Markdown files to HTML with Mermaid diagram support"
readme = "README.md"
authors = [{name = "Muly Benjamin", email = "muly.be@gmail.com"}]
license = {file = "LICENSE"}
keywords = ["markdown", "html", "converter", "mermaid"]
dependencies = [
    "markdown>=3.3",
    "python-frontmatter>=1.0.0",
    "pymdown-extensions>=9.0",
    "Pygments>=2.0.0"
]
requires-python = ">=3.11"

[project.urls]
Homepage = "https://github.com/samhakosem/markdown-to-html-converter"
Repository = "https://github.com/samhakosem/markdown-to-html-converter.git"

[project.scripts]
md2html = "md2html.converter:main"

[tool.setuptools]
packages = ["md2html"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
md2html = ["py.typed"]
  